调整测试方法，将mybatisPlus高级功能测试移动到同一个类中
