+++ main/src/test/resources/h2sql/initBiz.sql
@@ -6,6 +6,10 @@ INSERT INTO u_biz_role (id, name, name_zh, description, status, merchant_id, cre
                         update_user_id, update_time, deleted, version)
 VALUES (775413365652271104, 'ROLE_ADMIN', '管理员', '商家管理员', 1, 1, null, null, null, '2020-11-09 17:35:54', 0, 0);
 
+INSERT INTO u_biz_role (id, name, name_zh, description, status, merchant_id, create_user_id, create_time,
+                        update_user_id, update_time, deleted, version)
+VALUES (123413365652243104, 'ROLE_ADMIN', '管理员', '商家管理员', 1, 1, null, null, null, '2020-11-09 17:35:54', 0, 0);
+
 INSERT INTO u_biz_permission (id, parent_id, name, level, icon, type, url, method, status, merchant_id,
                               create_user_id, create_time, update_user_id, update_time, deleted, version)
 VALUES (775411429637042176, null, '用户管理', '1', null, null, '/bizUser/**', null, null, 1, null, null, null,
@@ -21,4 +25,5 @@ VALUES (775412006114766848, 'tom', '$2a$10$eqfBk4Py6KQuL1Vw4cxrq.xTcxBFUYBBNnCd4
         null, '张三', '2020-11-09 17:30:29', null, 1, null, null, null, '2020-11-09 17:30:29', 0, 0);
 
 INSERT INTO u_biz_user_role (id, user_id, role_id, merchant_id, create_user_id, create_time, update_user_id,
-update_time, deleted, version) VALUES (1, 775412006114766848, 775413365652271104, 1, null, null, null, null, 0, 0);
+                             update_time, deleted, version)
+VALUES (1, 775412006114766848, 775413365652271104, 1, null, null, null, null, 0, 0);
